test: separate project parse coverage
